2023.9.1: Change Log output for Screen time

+++ components/ehmtxv2/EHMTX_queue.cpp
@@ -92,72 +92,72 @@ namespace esphome
       ESP_LOGD(TAG, "empty slot");
       break;
     case MODE_BLANK:
-      ESP_LOGD(TAG, "queue: blank screen for %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: blank screen for %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_COLOR:
-      ESP_LOGD(TAG, "queue: color screen for %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: color screen for %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_CLOCK:
-      ESP_LOGD(TAG, "queue: clock for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: clock for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_DATE:
-      ESP_LOGD(TAG, "queue: date for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: date for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_GRAPH_SCREEN:
-      ESP_LOGD(TAG, "queue: graph for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: graph for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_FULL_SCREEN:
-      ESP_LOGD(TAG, "queue: full screen: \"%s\" for: %d sec", this->icon_name.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: full screen: \"%s\" for: %.1f sec", this->icon_name.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_ICON_SCREEN:
-      ESP_LOGD(TAG, "queue: icon screen: \"%s\" text: %s for: %d sec", this->icon_name.c_str(), this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: icon screen: \"%s\" text: %s for: %.1f sec", this->icon_name.c_str(), this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_ICON_PROGRESS:
-      ESP_LOGD(TAG, "queue: icon progress: \"%s\" text: %s for: %d sec", this->icon_name.c_str(), this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: icon progress: \"%s\" text: %s for: %.1f sec", this->icon_name.c_str(), this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_ICON_CLOCK:
-      ESP_LOGD(TAG, "queue: icon clock: \"%s\" for: %d sec", this->icon_name.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: icon clock: \"%s\" for: %.1f sec", this->icon_name.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_ICON_DATE:
-      ESP_LOGD(TAG, "queue: icon date: \"%s\" for: %d sec", this->icon_name.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: icon date: \"%s\" for: %.1f sec", this->icon_name.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_ALERT_SCREEN:
-      ESP_LOGD(TAG, "queue: icon: \"%s\" for: %d sec", this->icon_name.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: icon: \"%s\" for: %.1f sec", this->icon_name.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_TEXT_SCREEN:
-      ESP_LOGD(TAG, "queue: text text: \"%s\" for: %d sec", this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: text text: \"%s\" for: %.1f sec", this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_RAINBOW_ICON:
-      ESP_LOGD(TAG, "queue: rainbow icon: \"%s\" text: %s for: %d sec", this->icon_name.c_str(), this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: rainbow icon: \"%s\" text: %s for: %.1f sec", this->icon_name.c_str(), this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_RAINBOW_TEXT:
-      ESP_LOGD(TAG, "queue: rainbow text: \"%s\" for: %d sec", this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: rainbow text: \"%s\" for: %.1f sec", this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_RAINBOW_CLOCK:
-      ESP_LOGD(TAG, "queue: rainbow clock for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: rainbow clock for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_RAINBOW_DATE:
-      ESP_LOGD(TAG, "queue: rainbow date for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: rainbow date for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_ICON_TEXT_SCREEN:
-      ESP_LOGD(TAG, "queue: icon text screen: \"%s\" text: %s for: %d sec", this->icon_name.c_str(), this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: icon text screen: \"%s\" text: %s for: %.1f sec", this->icon_name.c_str(), this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_RAINBOW_ICON_TEXT_SCREEN:
-      ESP_LOGD(TAG, "queue: rainbow icon text screen: \"%s\" text: %s for: %d sec", this->icon_name.c_str(), this->text.c_str(), static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: rainbow icon text screen: \"%s\" text: %s for: %.1f sec", this->icon_name.c_str(), this->text.c_str(), this->screen_time_ / 1000.0));
       break;
     case MODE_FIRE:
-      ESP_LOGD(TAG, "queue: fire for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: fire for: %.1f sec", this->screen_time_ / 1000.0));
       break;
 
 #ifndef USE_ESP8266
     case MODE_BITMAP_SCREEN:
-      ESP_LOGD(TAG, "queue: bitmap for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: bitmap for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_BITMAP_SMALL:
-      ESP_LOGD(TAG, "queue: small bitmap for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: small bitmap for: %.1f sec", this->screen_time_ / 1000.0));
       break;
     case MODE_RAINBOW_BITMAP_SMALL:
-      ESP_LOGD(TAG, "queue: rainbow small bitmap for: %d sec", static_cast<uint16_t>(this->screen_time_ / 1000));
+      ESP_LOGD(TAG, "queue: rainbow small bitmap for: %.1f sec", this->screen_time_ / 1000.0));
       break;
 #endif
